Feature/reserves

+++ app/src/main/kotlin/com/itrocket/union/reserves/domain/entity/ReservesDomain.kt
@@ -2,6 +2,7 @@ package com.itrocket.union.reserves.domain.entity
 
 import android.os.Parcelable
 import com.itrocket.union.accountingObjects.domain.entity.ObjectInfoDomain
+import java.math.BigDecimal
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
@@ -10,6 +11,6 @@ data class ReservesDomain(
     val title: String,
     val isBarcode: Boolean,
     val listInfo: List<ObjectInfoDomain>,
-    val itemsCount: Int,
+    val itemsCount: Long,
     val comment: String = ""
 ) : Parcelable
